Makefile: Clean, search for libraries, dynamic/static compilation
Compile with the static .a libraries only when compiling the static
exe. Search for the .a static libraries in system automatically and
warn if they are not found. Clean the rules to use the library
paths found. Add dynamic libraries for boost and htslib when
linking the normal exe (the non static one).

Signed-off-by: Rick Wertenbroek <rick.wertenbroek@gmail.com>